1. Use env variable "EXA_BUILD_LOG" to set custom log path 2. Removed BaseTask.get_main_log_file()
